requirements-dev.txt
pyradiomics>=3.1
pytorch~=2.0
pyinstrument
SimpleITK
tqdm